+++ scripts/portable.mjs
@@ -34,9 +34,14 @@ async function resolvePortable() {
   if (!fs.existsSync(path.join(configDir, 'PORTABLE'))) {
     await fsp.writeFile(path.join(configDir, 'PORTABLE'), '')
   }
+  const tauriConf = JSON.parse(
+    await fsp.readFile('./src-tauri/tauri.conf.json', 'utf8'),
+  )
+  const appExe = `${tauriConf.productName}.exe`
+
   const zip = new AdmZip()
 
-  zip.addLocalFile(path.join(releaseDir, 'clash-verge.exe'))
+  zip.addLocalFile(path.join(releaseDir, appExe))
   zip.addLocalFile(path.join(releaseDir, 'verge-mihomo.exe'))
   zip.addLocalFile(path.join(releaseDir, 'verge-mihomo-alpha.exe'))
   zip.addLocalFolder(path.join(releaseDir, 'resources'), 'resources')
@@ -45,7 +50,7 @@ async function resolvePortable() {
   const require = createRequire(import.meta.url)
   const packageJson = require('../package.json')
   const { version } = packageJson
-  const zipFile = `Clash.Verge_${version}_${arch}_portable.zip`
+  const zipFile = `Clash.Verge.for.v2free_${version}_${arch}_portable.zip`
   zip.writeZip(zipFile)
   console.log('[INFO]: create portable zip successfully')
 }